requirements-dev.txt
black
piexif
pytest
pympler
Pillow
setuptools>=41.2.0
